+++ src/backend/admin/operations.ts
@@ -1,0 +1,16 @@
+import { ParameterizedContext } from 'koa';
+
+import { Database } from '../../db.js';
+
+export async function submitTimelineUpdateJobs(
+  ctx: ParameterizedContext,
+  db: Database
+): Promise<void> {
+  db.getTimelineUpdater()
+    ?.submitUpdateJobs()
+    .then((n) => n)
+    .catch((e) => e);
+  ctx.body = 'update process started';
+  ctx.type = 'text';
+  ctx.status = 200;
+}
